Allow uploading files bigger than 64k

+++ src/lib/surl.h
@@ -2,7 +2,7 @@
 #define __surl_h
 
 #include <arpa/inet.h>
-
+#include "platform.h"
 #include "../surl-server/surl_protocol.h"
 #include "simple_serial.h"
 
@@ -50,7 +50,7 @@ size_t __fastcall__ surl_receive_lines(char *buffer, size_t max_len);
 void surl_strip_html(char strip_level);
 void surl_translit(char *charset);
 
-int __fastcall__ surl_send_data_params(size_t total, int raw);
+int __fastcall__ surl_send_data_params(uint32 total, int raw);
 #define surl_send_data(buffer, len) simple_serial_write(buffer, len)
 
 int __fastcall__ surl_find_line(char *buffer, size_t max_len, char *search_str);
