+++ src/mastodon/api/oauth.c
@@ -197,7 +197,7 @@ password_again:
     dputs("POST "LOGIN_URL"... ");
     resp = surl_start_request(SURL_METHOD_POST, login_url, NULL, 0);
 
-    surl_send_data_params(post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
+    surl_send_data_params((uint32)post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
     surl_send_data(post, post_len);
     free(post);
 
@@ -243,7 +243,7 @@ otp_again:
       dputs("POST "LOGIN_URL"... ");
       resp = surl_start_request(SURL_METHOD_POST, login_url, NULL, 0);
 
-      surl_send_data_params(post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
+      surl_send_data_params((uint32)post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
       surl_send_data(post, post_len);
       free(post);
 
@@ -287,7 +287,7 @@ otp_again:
     dputs("POST "OAUTH_URL"... ");
     resp = surl_start_request(SURL_METHOD_POST, oauth_url, NULL, 0);
 
-    surl_send_data_params(post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
+    surl_send_data_params((uint32)post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
     surl_send_data(post, post_len);
     free(post);
 
@@ -356,7 +356,7 @@ int register_app(void) {
   resp = surl_start_request(SURL_METHOD_POST, reg_url, NULL, 0);
   free(reg_url);
 
-  surl_send_data_params(post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
+  surl_send_data_params((uint32)post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
   surl_send_data(post, post_len);
   free(post);
 
@@ -430,7 +430,7 @@ int get_oauth_token(void) {
 
   post = prepare_oauth_token_post();
   post_len = strlen(post);
-  surl_send_data_params(post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
+  surl_send_data_params((uint32)post_len, SURL_DATA_X_WWW_FORM_URLENCODED_HELP);
   surl_send_data(post, post_len);
   free(post);
